test: Add an integration test for RemoteSettings reconfigure
