+++ merino-integration-tests/src/reconfigure/mod.rs
@@ -1,0 +1,2 @@
+mod redis;
+mod remote_settings;
